+++ changelog.md
@@ -3,6 +3,8 @@
 ## 0.8.31 
 
 - Bug fix multilne journal enrich html render
+- Add two new adavanced options for event link `onlyIfHidden`,`onlyIfUnHidden` for execute the effects only if the lement stat hidden is tru or false (for now tested only with token and drawing)
+- Add new module setting 'Disable all trigger for hidden element'
 
 ## 0.8.30